refactor(browser): use ../utils
